refactor: 重构接口安全功能和常量到utils中
