+++ src/renderer/utils/security/data/index.ts
@@ -1,0 +1,7 @@
+export const BASE32_ALPHABET = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ234567';
+
+export const DEFAULT_TOTP_DIGITS = 6;
+
+export const DEFAULT_TOTP_PERIOD_SECONDS = 30;
+
+export const DEFAULT_TOTP_HMAC_HASH = 'SHA-1';
